# rcb: prerequisites for GPU acceleration