This PR changes the "array of structs" data structure in RCB to a "struct of arrays" (one for the weights, one for part IDs, and one for each dimension) to provide efficient data transfers to the GPU (and also allow for SIMD processing)

TODO
- [x] update commented-out unit tests