@hhirtz 2022-06-20
2D triangular mesh, 20 million triangles

RCB, 1 iteration (2 parts), sequential times:
- master: 1.46 s
- this pr (gpuprep): 1.28 s

RCB, 12 iterations (4096 parts), sequential times:
- master: 8.24 s
- this pr (gpuprep): 6.23 s

![base strong](https://user-images.githubusercontent.com/51088794/174611727-246b6b68-c4d1-4ac7-ab43-f0a49351b206.svg)