@hhirtz 2022-06-23
Thanks to #170 and #175, weak scaling measurements can be performed.

3D triangular mesh, starting from 2.18 millions of triangles for 1 thread, to 559 millions for 256 threads.
 
RCB, 1 iteration (2 parts), sequential times:

- master: 171 ms
- this pr (gpuprep): 96.8 ms

RCB, 12 iterations (4096 parts), sequential times:

- master: 4.46 s
- this pr (gpuprep): 393 ms

![base weak](https://user-images.githubusercontent.com/51088794/175288778-94f86e56-cd15-420f-9a00-12940cbf10a9.svg)

`Speedup = N * walltime(1 thread) / walltime(N threads)`